新的分支

+++ last_1/src/main/java/readNote/service/ReadNoteService.java
@@ -29,6 +29,6 @@ public interface ReadNoteService {
 	 * @param id
 	 * @return
 	 */
-	public ReadNote getReadNote(Integer id);
+	public ReadNote getReadNote(Long id);
 	
 }
